Bootstraps user sign in endpoint.
This commit installs several gems that we are going to use in the development like rspec, devise, cors and other testing gems. It also covers configuration, the addition of helpers in the test suite that are going to become handy for api requests testing such as response json method.

There is more routing config that needs to be done in the upcoming commits for the resources endpoints. Moving the devise endpoint inside the api namespace may be a good choice as well to keep consistency in the api urls.